+++ options_test.go
@@ -15,11 +15,47 @@
 package goalesce
 
 import (
+	"errors"
 	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
+
+var (
+	withMockDeepCopyError Option = func(c *coalescer) {
+		c.deepCopy = func(v reflect.Value) (reflect.Value, error) {
+			return reflect.Value{}, errors.New("mock DeepCopy error")
+		}
+	}
+	withMockDeepMergeError Option = func(c *coalescer) {
+		c.deepMerge = func(v1, v2 reflect.Value) (reflect.Value, error) {
+			return reflect.Value{}, errors.New("mock DeepMerge error")
+		}
+	}
+)
+
+func withMockDeepCopyErrorWhen(expected interface{}) Option {
+	return func(c *coalescer) {
+		c.deepCopy = func(v reflect.Value) (reflect.Value, error) {
+			if expected == v.Interface() {
+				return reflect.Value{}, errors.New("mock DeepCopy error")
+			}
+			return c.defaultDeepCopy(v)
+		}
+	}
+}
+
+func withMockDeepMergeErrorWhen(expected1, expected2 interface{}) Option {
+	return func(c *coalescer) {
+		c.deepMerge = func(v1, v2 reflect.Value) (reflect.Value, error) {
+			if expected1 == v1.Interface() && expected2 == v2.Interface() {
+				return reflect.Value{}, errors.New("mock DeepMerge error")
+			}
+			return c.defaultDeepMerge(v1, v2)
+		}
+	}
+}
 
 func TestWithTypeCopier(t *testing.T) {
 	called := false
